Re-Chunkify
Made a new implementation for my chunkify function. I was using the one I already made over in Gamedata-Parser, and I decided to rewrite it using a Generator function. Inspired from this StackOverflow answer, it works really well (https://stackoverflow.com/questions/8495687/split-array-into-chunks/55435856#55435856).

Started working on this a few days ago, I think it was October 3, the nighttime/early morning.

This was going to be part of a few other changes also, but I wasn't sure if those were quite polished enough to be ready to push. About to work on those after this commit!