Aquatic Chunks! v0
With more thanks to UtterEvergreen1, they decompiled the source code for MCCToolChest, and have been working with the chunk parsing bits, and made it work for Aquatic+ versions! Since I'm not familiar with working in C++ (I can read it now, just no experience in writing it), I rewrote the source the best I could, over in TypeScript. This is more to try and reverse engineer how the format works, using the source, then to get a working parser the first time around. This version will definitely not be able to parse anything. Now that I rewrote it in something more familiar to me, I can start abstracting simple parts down so it is easier to understand. That's what I did with NBT.js, and it worked very well. Take something big, then tear it down, and put it back together to look more like you want it to. Eventually, things start to click, and you will understand the implementation enough to write your own program for the format, and then you can document the format using that! So, cheers to the last 5 hours or so of trying to understand the reverse-engineered MCCToolChest code, but in TypeScript! haha.

Oh yeah, one other note:
I didn't convert the functions that weren't used by `class AquaticParser` itself, like `method_3`, `method_9`, `method_11`, and a few others.