Chunk Location Offset Implementation
Now, on to the app implementation of the bug fix! It actually appears to be working so far!

I was calculating the chunk locations offset section wrong, and this was throwing off the rest of the reading process. So, turns out it was the `Location.offset` calculation that was incorrect, rather than the chunk reading part.

In part of writing the new code that fixes this issue, I modernized some of the Region class, removing some internal methods that don't need to be there anymore. I essentially moved `#readLocation()` directly into `readLocations()`, with the new edits though. Now it's both smaller, more compact, and it actually seems to be reading correctly, so far! Yay. Been hanging on this issue for a few months now, and it has been holding me back from wanting to work on the project. I'm still essentially unfamiliar with how to use bitwise operators, and I think that's what messed me up. I was making the calculations wrong.