Chunk Structure Header
Wow! It's working! Even more thanks for expert help from UtterEvergreen, now I can decompress all of the chunks from region files. It's been hard to figure out how to calculate the correct values for the chunk header declarations. I think it's because I haven't had much experience with bitwise operators. It's also probably just because working with this data is kind of complex. I'm up for that though! The biggest thing I've learned is that if it's possible to read the data, then I'm gonna figure out how to do it eventually too! It just may take some time :)

https://github.com/UtterEvergreen1/Minecraft-LCE-Docs/blob/main/Documentation.md#chunk-structure