Corner Checking
Now I'm at the part of the code where I want to check all corners of it, so nothing can slip by it's structure.

This one has a bit more renaming, and things like that. Looking into how I did the RLE decoder function, I'm going to refactor that a bit, to make it similar to how the NBTify `NBTWriter` works, in the way that it can auto-allocate any bytes it needs, on the fly. The RLE decoder currently requires you to know how big it is on the inflate process.

Listening to King's X Dogman now! I really like 'Sunshine Rain' :)