ChunkData Header Setup
Starting to work on reading the chunk data itself! I'm renaming the current header parsing logic to be specifically the compression part. There's gonna be two stages, the decompression header stage, which allows the rest of the chunk data to be decompressed, and then the parts after that, which will have the chunk data itself!

https://github.com/UtterEvergreen1/Minecraft-LCE-Docs/blob/main/ChunkData.md

This is the committing part of this code, so I'm adding a few changes as I commit them to help with the debugging process, since I ended up finding those a few edits later in. In the test file, I'm reading a chunk at a given offset, as I want to read one that is an Aquatic version 12 one. The first chunk in the region that I was reading from as of now has been a version 11 one, and that has a different internal format than version 12. So I am picking a random chunk in the region that *is* version 12, just to not cause myself any goofy errors, from reading the wrong kind of chunk unexpectedly.