NBT Search
Now that NBTify is extra stable, with things like it's reading checksum and format auto-detection, I set up a little script to try and find any NBT inside of the chunk. No luck! I'm curious what the text snippets at the end of the chunk are. You can see them if you stringify the bytes from the chunk. That's a commented out part of this test script too (the `Buffer.toString()` part), if you wanted to see what that looks like.

This code was originally from way back when I wrote that `anvil.mjs` file, back in October it looks like. I just modified it a bit and put it back into the test script for the current version of the library.