RLE Rework
Working on some more debugging with UtterEvergreen, and found out my current RLE decoding implementation is incorrect, so only some values of the data managed to be decompressed as expected. There's still some random ghost data at the end of the buffer for some reason, but this new rewrite makes the first part of the data finally return correct! Commiting it now to keep it safe :)