RLE Debug
When working with UtterEvergreen to debug my RLE implementation, turns out the extra length at the end of my result buffer was actually there as expected! His file export example was actually saving in a smaller file than it was supposed to. It was a missing `fwrite` call for the demo, and that made it save in the same amount of bytes as the original, compressed file, which was 8196 bytes long.