Blocks Section Parsing: Part One (Three)
Been working on this off and on all day! While trying to add this part of the logic, I encountered all of those errors with the RLE code of mine, so I had to come back to this only once that got fixed first.

I tried a few different ways to keep track of the `DataView` byte offset state, and this one seems to be the best, when it has to be dynamic/is too hard to keep track of yourself in the code. This makes it work like how NBTify handles the byte offset state, by keeping it in a variable, and updating the offset variable after you read the data as needed.

I wish I didn't have to do the extra subtraction after doing the byte offset assignment, but returning `byteOffset += *num*` will return the new value of `byteOffset`, while when using it to access the `DataView`, I want the original value. So, to do the assignment and access the data at the correct offset, I have to subtract the new offset change from that last expression. Kind of messy, but so far I do tend to like them on one line. It would be neat if `DataView` could automatically keep a pointer-like offset like that for you, so you didn't have to manage it yourself. And if you could modify it as a property from the outside too, then it would essentially solve both problems. Something like this, maybe:

```ts
declare interface DataView {
  /**
   * Internal state management to keep track of where to read and write to and from the source ArrayBuffer at.
  */
  pointer: number;
}

declare interface DataViewGetOptions {
  /**
   * Defaults to false.
  */
  littleEndian?: boolean;
  /**
   * Defaults to false. When true, this will increment the pointer attribute on the DataView by the byte length amount read or written by the given read or write call.
  */
  pointer?: boolean;
}

declare const view: DataView & {
  getUint8(byteOffset: number, littleEndian?: boolean): number;
  getUint8(byteOffset: number, options?: DataViewGetOptions): number;
};

// Not real config options, just an idea
// This will increment the `DataView.prototype.pointer` property by 1, because it was called by `getUint8`.
const aRandomInt = view.getUint8(view.pointer,{ pointer: true });
```