Deflate-Raw, Nearly Browser
While I really like the structure of the Compression Streams API, it is a bit of a hassle to work with streams all of the time. I wish I could just use them as async functions that I could chain together with my Uint8Array logic. There doesn't seem to be a simple way to stay directly within the ArrayBuffer/Uint8Array system without going to streams to compress and decompress things. I'm not sure whether I should make a small library to smoothen this out, which all of my projects can use, or if I should just implement these two functions on my own in every project individually. Either way, I don't really like either option :(

I do think it was a good choice to force it to be async though, that could cause problems if it was a really big file to work with.

Starting fresh with this codebase too, so I removed the old 'chunk' demo test from a while back. I'm going to read the format closer again, and see what my new eyes can do instead of reimplementing what I already have. That seemed to help a lot over in Gamedata-Parser. Then I only have to debug those new edits against how this current codebase stands if the new one isn't working right (that's what happened over in Gamedata-Parser, too). Listening to 'Glop', on 'You Must Be This Tall'. Bought 4 albums today XD; Wing Beat Fantastic, You Must Be This Tall, Plague Accommodations, and Marco's new album Their Colors Fade. Going to listen to that for the first time through on the way to work today :)

'Glop' sounds a lot like 'Salve-Dependent Scorpions', and 'Evidence Of Humanity' as a whole. This sound is really cool to me. I think it's making it hard for me to write "regular" songs at the moment now though, haha.