Rework Take One
Listening to Their Colors Fade, couldn't wait until the drive to work :)

Unless my region files are weird, I think my chunk locations parsing might still not be working right. There's a few blocks of 5 chunks that shows up in every few tens of chunks in the locations list. This appeared to be the same for the timestamp implementation I wrote (then removed, before this commit). And looking back at the implementation before this initial rework, it was doing that before, too. So I'm not sure if everything is alright or not quite yet. Going to try implementing things some more before digging too deep into it.

I initially had two functions to read the locations, just like the old implementation and Gamedata-Parser. But I realized that it would be simpler to just iterate over the dedicated location bytes once, and only have to create a single DataView to read all of them. I may change over the Gamedata-Parser definitions part to also do it like this, as I think it could remove some of the extra garbage collection steps. Thanks, ThePrimeagen :P