Chunk Timestamps
Had a working version for this the other day. I wasn't sure if the values being read were correct or not though, so I held off on adding that version. I think these still aren't in the right format to be usable timestamps, but I think I am parsing the data right though, so I'm adding this implementation logic now that I am more sure that it is right most of the way.

I'm getting help from UtterEvergreen1 to figure out how to convert the values as they are to a usable date that does reflect the actual times represented by each timestamp.