Chunk NBT Discovery
Using some modern buffer traversal logic in conjunction with NBTify, I found out how to discover where the NBT is located at the end of each chunk buffer section!

I realized that looping from the end of the buffer to start with could probably be faster, since the NBT is likely near the end of the buffer, and that made a big difference! Super neat improvement.

(Thanks, Stack-O :))
https://stackoverflow.com/questions/15743805/how-to-decrement-work-in-javascript-for-loop

Found a bug in NBTify while working on this, it had to do with the automatic format detection logic, specifically the BedrockLevel handling. Even with stating that BedrockLevel should be off, it was still showing what bytes would be a BedrockLevel header, had it actually been a BedrockLevel file.

Looking into the LCE-Docs again, as I want to get progress on this format going again. Things are really close to being super cool, and the formats just have to be readable from an implementation perspective. I pretty much know how I want the API to look, work, and feel at this point. Lots of cool things :)