ReadonlyArray Regions
Learned about Object.seal(), and how it can prevent the addition of new properties on a new object! Thinking about it, this works great for Region-representing arrays! It can prevent you from expanding the length of the array, which works really well here because Regions can only hold 1024 chunks. When generating a new Region array, the values can be pre-filled with `null`, since I already use that to represent ungenerated Chunks. It doesn't prevent the editing of existing properties (yay), it only prevents the addition of new properties. I discovered this with the difference of `Object.seal(Array(25))` compared to `Object.seal(Array(25).fill(null))`. Only the second one allows you to fill in the values, even with the Array length being specified ahead of time.  Knowing that Arrays are just Objects in JS helps a lot, since Array items are just keys on the 'Object', in the conceptual sense, as well as the literal sense.

I was going to just use the index-signature by itself, but then I realized that it essentially will work like the `ReadonlyArray` type, with the only difference being that array items should be allowed to be reassignable, which they aren't with the base `ReadonlyArray` type.

https://stackoverflow.com/questions/21988909/is-it-possible-to-create-a-fixed-length-array-in-javascript
https://www.google.com/search?q=object.seal+vs+object.freeze