SizeOfSubChunks + Validation Demo
The next test of the parsing, as well as the demo I mentioned in the previous commit. Wanting to see if the data coming out of the parser looks correct so far, as whether that's the case will determine if further properties read after these will also be read correctly.