Allocation Error Fix!
Wow, what a goofball. Pushing to an `Array(length)` just increases the size of the array. So if you're pushing with every loop, and the loop is based on the length of the array, you can see how that goes, how that goes, how that goes, how that goes, haha. That's some smart maneuvers right there. I like that, definitely makes 'a lot of sense'. Thinking of it in JS terms though, it actually really does make sense. I think I missed it because I was kind of thinking of the `Array` like how the `ArrayBuffer`-like objects behave, that the length can't be changed. But I tend to just use `.push()` for most things, so I think that's why it wen't under the radar. And I also used the length of the array for the loop instead of just 4 because I didn't want to use redundant magic values. Might as well just declare them in a single place, at least I tend to try doing that. Sometimes that's not ideal though.