Seeking SectionJumpTable
The parsing isn't working correctly, and I think it's because I forgot a few cases of moving the pointer of the parser's `DataView` when reading values. Looks like it's getting closer to finding the rest of the problems though!

Trying to move the use of some magic numbers into singularly-defined places, to make it more apparent as to what they are for/how they are used.