MaxSectionAddress
This is a summary between this commit and the previous one.

Did some debugging with Evergreen, and we figured out what parts might be/were causing the parsing errors.

The first part was that I forgot to update the reading offset in the buffer after reading the values from the `sectionJumpTable`.

~~And the current finding I still have to resolve is that `seek()` in his implementation is absolute to the start of the buffer, rather than additive to how many times you call it/with the length you want to move forward in the buffer by~~ (`rewind()`)

And also it turns out that the `76 + address` calculation is `76` for all of the bytes for all of the values before that point in the parse stage, and address is also absolute to the start of the buffer, rather than relative to where you previously just were in the `sectionJumpTable` loop.

Ok, and for the changes specific to this commit, Evergreen noticed my `address` reading code wasn't reading the values from the proper offset, meaning the next `address` would read bytes from the previous value's `address` position.