Region Refactorings
Now the Region Entry headers are read during the Region stage, this is just like how I did it in MCA-Buffer recently.

The demo currently logs only the single Chunk I've been testing on, but it's parsing the whole Region, as I want to test as much as possible for any errors caused by anything I may have written wrong or missed on accident.