Readx128 Debugging
This has something to do with the dual-pointer implementation that this reader needs to have, but doesn't currently have to function properly.

On line 150 you can see the demo code I wrote which allows the rest of the parsing to run, whether or not the values are inherently correct. I'm pretty sure they're not, I just kind of wanted to try to see if parsing as Little Endian was the way to go instead, and I think it just happened to work based on the offsets it read/hallucinated.

Also added temporary types and commented existing parameters from the original code so there are less errors from those placeholders.