DataInputManager Migration
Going to try using/making a fork of Evergreen's buffer reader since I think most of my errors are around remembering when to move the pointer forward on the DataView. I thought I could figure that out myself without having to rebuild Evergreen's DataInputManager reader, but looks like that may have to be the next step to figure out where to increment the pointer at the correct spots.