SectionJumpTable Validation
Trying to debug this more with Evergreen, he said that the values for the current reading of the SectionJumpTable appear to be correct. He also said that his code does Little Endian parsing, which I didn't know beforehand. So I'm going to go back and try parsing with those in some spots maybe.